Rename config class file, working on README.
